feat: add leaflet-show option to help with hidden maps
The idea is to help with hiddden place like with ng-show/ng-hide div. What you have to know is if
your map has been initialized in an hidden place it will be broken once you display the container.
You must fix it and this option help you for that. Put an expression that is evaluated to true when
the container is visible.